move 4w tests
